### 2024-02-29 #6432 @editorialbot
[PRE REVIEW]: DynamicOED.jl: A Julia package for solving optimum1 experimental design problems
**Submitting author:** @AlCap23 (<a href="http://orcid.org/0000-0003-4143-3040">Julius Martensen</a>)
**Repository:** https://github.com/mathopt/DynamicOED.jl
**Branch with paper.md** (empty if default branch):
**Version:** v.0.1.1
