### 2024-02-29 #6432 @danielskatz
Actually, as the non-testing portion of this is about 1000 LOC, I want the editors to discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS before I look for an editor. You should hear back in a week or two.
