### 2024-03-01 f6fd7fb
CompatHelper: bump compat for Optimization to 3, (keep existing compat)
