+++ Project.toml
@@ -21,7 +21,7 @@ CommonSolve = "0.2"
 ComponentArrays = "0.15"
 DocStringExtensions = "0.9"
 ModelingToolkit = "<8.73.0"
-Optimization = "<3.20.0"
+Optimization = "<3.20.0, 3"
 OrdinaryDiffEq = "6"
 Reexport = "1.2"
 SciMLBase = "<2.8.0"
