### 2024-03-13 c2adece
CompatHelper: bump compat for SciMLBase to 2, (keep existing compat)
