+++ Project.toml
@@ -24,7 +24,7 @@ ModelingToolkit = "<8.73.0"
 Optimization = "<3.20.0"
 OrdinaryDiffEq = "6"
 Reexport = "1.2"
-SciMLBase = "<2.8.0"
+SciMLBase = "<2.8.0, 2"
 SciMLSensitivity = "7"
 julia = "1.6"
 
