### 2024-03-15 #6432 @AlCap23
Hey!
First of all, thanks! I am terribly sorry for being dormant this last few weeks. I had some in person workshops and deadlines coming up.

I'll get to the references ASAP. Additionally one of our dependencies had a major (breaking) release and I still need to figure out the right configuration of bounds to get the current stable version registered before the release.

Feel free to reach out any time ( either here or via email ).

Cheers,

Julius
