+++ test/criteria.jl
@@ -13,10 +13,10 @@ end
 
 @testset "Criteria" begin
     F = [5.68145 -0.148312 2.00382 5.1841 3.57814
-        -0.148312 2.17276 -1.18576 1.57651 -3.95544
-        2.00382 -1.18576 3.8703 -0.14832 2.69545
-        5.1841 1.57651 -0.14832 10.9617 0.0774284
-        3.57814 -3.95544 2.69545 0.0774284 24.8495]
+         -0.148312 2.17276 -1.18576 1.57651 -3.95544
+         2.00382 -1.18576 3.8703 -0.14832 2.69545
+         5.1841 1.57651 -0.14832 10.9617 0.0774284
+         3.57814 -3.95544 2.69545 0.0774284 24.8495]
 
     test_results = [(FisherACriterion(), -47.53571),
         (FisherDCriterion(), -2180.4107236837117),
