+++ Project.toml
@@ -20,11 +20,11 @@ Aqua = "0.7"
 CommonSolve = "0.2"
 ComponentArrays = "0.15"
 DocStringExtensions = "0.9"
-ModelingToolkit = "<8.73.0"
-Optimization = "<3.20.0"
+ModelingToolkit = "~8.72.0"
+Optimization = "~3.19.0"
 OrdinaryDiffEq = "6"
 Reexport = "1.2"
-SciMLBase = "<2.8.0"
+SciMLBase = "~2.7.0"
 SciMLSensitivity = "7"
 julia = "1.6"
 
