+++ paper/bibliography.bib
@@ -8,15 +8,16 @@
 }
 
 @article{bezanson2017julia,
-  title={Julia: A fresh approach to numerical computing},
-  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
-  journal={SIAM review},
-  volume={59},
-  number={1},
-  pages={65--98},
-  year={2017},
-  publisher={SIAM},
-  url={https://doi.org/10.1137/141000671}
+author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+title = {Julia: A Fresh Approach to Numerical Computing},
+journal = {SIAM Review},
+volume = {59},
+number = {1},
+pages = {65-98},
+year = {2017},
+doi = {10.1137/141000671},
+URL = {https://doi.org/10.1137/141000671},
+eprint = {https://doi.org/10.1137/141000671}
 }
 
 @software{vaibhav_kumar_dixit_2023_7738525,
@@ -104,12 +105,18 @@
   keywords = {Automatic differentiation,Differential\textendash algebraic equations,Sensitivity analysis},
 }
 
-@inproceedings{juniper,
-     Author = {Ole Kröger and Carleton Coffrin and Hassan Hijazi and Harsha Nagarajan},
-     Title = {Juniper: An Open-Source Nonlinear Branch-and-Bound Solver in Julia},
-     booktitle="Integration of Constraint Programming, Artificial Intelligence, and Operations Research",
-     pages="377--386",
-     year="2018",
-     publisher="Springer International Publishing",
-     isbn="978-3-319-93031-2"
+@InProceedings{juniper,
+author="Kr{\"o}ger, Ole
+and Coffrin, Carleton
+and Hijazi, Hassan
+and Nagarajan, Harsha",
+editor="van Hoeve, Willem-Jan",
+title="Juniper: An Open-Source Nonlinear Branch-and-Bound Solver in Julia",
+booktitle="Integration of Constraint Programming, Artificial Intelligence, and Operations Research",
+year="2018",
+publisher="Springer International Publishing",
+address="Cham",
+pages="377--386",
+abstract="Nonconvex mixed-integer nonlinear programs (MINLPs) represent a challenging class of optimization problems that often arise in engineering and scientific applications. Because of nonconvexities, these programs are typically solved with global optimization algorithms, which have limited scalability. However, nonlinear branch-and-bound has recently been shown to be an effective heuristic for quickly finding high-quality solutions to large-scale nonconvex MINLPs, such as those arising in infrastructure network optimization. This work proposes Juniper, a Julia-based open-source solver for nonlinear branch-and-bound. Leveraging the high-level Julia programming language makes it easy to modify Juniper's algorithm and explore extensions, such as branching heuristics, feasibility pumps, and parallelization. Detailed numerical experiments demonstrate that the initial release of Juniper is comparable with other nonlinear branch-and-bound solvers, such as Bonmin, Minotaur, and Knitro, illustrating that Juniper provides a strong foundation for further exploration in utilizing nonlinear branch-and-bound algorithms as heuristics for nonconvex MINLPs.",
+isbn="978-3-319-93031-2"
 }
