+++ previews/PR24/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"api/#API","page":"API","title":"API","text":"","category":"section"},{"location":"api/#Criteria","page":"API","title":"Criteria","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"ACriterion\nDCriterion\nECriterion\nFisherACriterion\nFisherDCriterion\nFisherECriterion","category":"page"},{"location":"api/#DynamicOED.ACriterion","page":"API","title":"DynamicOED.ACriterion","text":"struct ACriterion <: DynamicOED.AbstractInformationCriterion\n\nThe A-Criterion for experimental design.\n\ntr(inv(F))\n\n\n\n\n\n","category":"type"},{"location":"api/#DynamicOED.DCriterion","page":"API","title":"DynamicOED.DCriterion","text":"struct DCriterion <: DynamicOED.AbstractInformationCriterion\n\nThe D-Criterion for experimental design.\n\ndet(inv(F))\n\n\n\n\n\n","category":"type"},{"location":"api/#DynamicOED.ECriterion","page":"API","title":"DynamicOED.ECriterion","text":"struct ECriterion <: DynamicOED.AbstractInformationCriterion\n\nThe E-Criterion for experimental design.\n\nmax(eigvals(F))\n\n\n\n\n\n","category":"type"},{"location":"api/#DynamicOED.FisherACriterion","page":"API","title":"DynamicOED.FisherACriterion","text":"struct FisherACriterion <: DynamicOED.AbstractInformationCriterion\n\nThe Fisher A-Criterion for experimental design.\n\n-tr(F)\n\n\n\n\n\n","category":"type"},{"location":"api/#DynamicOED.FisherDCriterion","page":"API","title":"DynamicOED.FisherDCriterion","text":"struct FisherDCriterion <: DynamicOED.AbstractInformationCriterion\n\nThe Fisher D-Criterion for experimental design.\n\n-det(F)\n\n\n\n\n\n","category":"type"},{"location":"api/#DynamicOED.FisherECriterion","page":"API","title":"DynamicOED.FisherECriterion","text":"struct FisherECriterion <: DynamicOED.AbstractInformationCriterion\n\nThe Fisher E-Criterion for experimental design.\n\n-min(eigvals(F))\n\n\n\n\n\n","category":"type"},{"location":"api/#Problem","page":"API","title":"Problem","text":"","category":"section"},{"location":"api/","page":"API","title":"API","text":"DynamicOED.OEDProblem\nDynamicOED.Timegrid","category":"page"},{"location":"api/#DynamicOED.OEDProblem","page":"API","title":"DynamicOED.OEDProblem","text":"struct OEDProblem{S, O, T, A, DO}\n\nThe basic definition of an optimal experimental design problem.\n\nFields\n\nsystem: The optimal experimental design system in form of an ODESystem\nobjective: The objective criterion\ntimegrid: The time grid\nalg: Solver for the differential equations\ndiffeq_options: Differential equations options\n\n\n\n\n\n","category":"type"},{"location":"api/#DynamicOED.Timegrid","page":"API","title":"DynamicOED.Timegrid","text":"struct Timegrid{V, I, G, T}\n\nA structure for holding a multi-variable time grid. \n\nFields\n\nvariables: The variables\nindicators: The indicator for switching variables\ntimegrids: The individual time grids\ntimespans: The overall time grid\n\n\n\n\n\n","category":"type"},{"location":"api/#Symbolic-Differentiation-Backends","page":"API","title":"Symbolic Differentiation Backends","text":"","category":"section"},{"location":"api/","page":"API","title":"API","t
[…]
