+++ paper/bibliography.bib
@@ -13,7 +13,6 @@
   doi = {10.1137/141000671},
   urldate = {2024-03-15},
   abstract = {JuMP is an open-source modeling language that allows users to express a wide range of optimization problems (linear, mixed-integer, quadratic, conic-quadratic, semidefinite, and nonlinear) in a high-level, algebraic syntax. JuMP takes advantage of advanced features of the Julia programming language to offer unique functionality while achieving performance on par with commercial modeling tools for standard tasks. In this work we will provide benchmarks, present the novel aspects of the implementation, and discuss how JuMP can be extended to new problem classes and composed with state-of-the-art tools for visualization and interactivity.},
-  file = {/Users/juliusmartensen/Library/Mobile Documents/iCloud~com~logseq~logseq/Documents/HilbertsHotel/assets/Bezanson et al. - 2017 - Julia A Fresh Approach to Numerical Computing.pdf}
 }
 
 @article{rackauckas2017,
@@ -25,7 +24,6 @@
   number = {1},
   doi = {10.5334/jors.151},
   keywords = {Applied Mathematics},
-  file = {/Users/juliusmartensen/Library/Mobile Documents/iCloud~com~logseq~logseq/Documents/HilbertsHotel/assets/Rackauckas and Nie - 2017 - DifferentialEquations.jl – a performant and featur.pdf}
 }
 
 @misc{vaibhav_kumar_dixit_2023_7738525,
@@ -35,7 +33,6 @@
   month = mar,
   doi = {10.5281/zenodo.7738525},
   howpublished = {Zenodo},
-  file = {/Users/juliusmartensen/Zotero/storage/C3YX8D8Z/7738525.html}
 }
 
 @phdthesis{Koerkel2002,
@@ -46,7 +43,6 @@
   doi = {10.11588/heidok.00002980},
   school = {Universit{\"a}t Heidelberg},
   keywords = {agbock experimental design DAE},
-  file = {/Users/juliusmartensen/Zotero/storage/ZPFDJSCT/Körkel - 2002 - Numerische Methoden für Optimale Versuchsplanungsp.pdf;/Users/juliusmartensen/Zotero/storage/CTAT9RVX/2980.html}
 }
 
 @inproceedings{juniper,
@@ -63,7 +59,6 @@
   abstract = {Nonconvex mixed-integer nonlinear programs (MINLPs) represent a challenging class of optimization problems that often arise in engineering and scientific applications. Because of nonconvexities, these programs are typically solved with global optimization algorithms, which have limited scalability. However, nonlinear branch-and-bound has recently been shown to be an effective heuristic for quickly finding high-quality solutions to large-scale nonconvex MINLPs, such as those arising in infrastructure network optimization. This work proposes Juniper, a Julia-based open-source solver for nonlinear branch-and-bound. Leveraging the high-level Julia programming language makes it easy to modify Juniper's algorithm and explore extensions, such as branching heuristics, feasibility pumps, and parallelization. Detailed numerical experiments demonstrate that the initial release of Juniper is comparable with other nonlinear branch-and-bound solvers, such as Bonmin, Minotaur, and Knitro, illustrating that Juniper provides a strong foundation for further exploration in utilizing nonlinear branch-and-bound algorithms as heuristics for nonconvex MINLPs.},
   isbn = {978-3-319-93031-2},
   langid = {english},
-  file = {/Users/juliusmartensen/Library/Mobile Documents/iCloud~com~logseq~logseq/Documents/HilbertsHotel/assets/Kröger et al. - 2018 - Juniper An Open-Source Nonlinear Branch-and-Bound.pdf}
 }
 
 @article{Li2000SensitivityAnalysisDifferential,
@@ -82,7 +77,6 @@
[…]
