### 2024-03-15 d0731e0
Paper Update II (#25)
* Update bib

* Update bib and add first name

* Rmv file field
