+++ paper/bibliography.bib
@@ -1,95 +1,69 @@
-@misc{ma2021modelingtoolkit,
-      title={ModelingToolkit: A Composable Graph Transformation System For Equation-Based Modeling},
-      author={Yingbo Ma and Shashi Gowda and Ranjan Anantharaman and Chris Laughman and Viral Shah and Chris Rackauckas},
-      year={2021},
-      eprint={2103.05244},
-      archivePrefix={arXiv},
-      primaryClass={cs.MS}
-}
-
 @article{bezanson2017julia,
-author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
-title = {Julia: A Fresh Approach to Numerical Computing},
-journal = {SIAM Review},
-volume = {59},
-number = {1},
-pages = {65-98},
-year = {2017},
-doi = {10.1137/141000671},
-URL = {https://doi.org/10.1137/141000671},
-eprint = {https://doi.org/10.1137/141000671}
+  title = {Julia: {{A Fresh Approach}} to {{Numerical Computing}}},
+  shorttitle = {Julia},
+  author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+  year = {2017},
+  month = jan,
+  journal = {SIAM Rev.},
+  volume = {59},
+  number = {1},
+  pages = {65--98},
+  publisher = {{Society for Industrial and Applied Mathematics}},
+  issn = {0036-1445},
+  doi = {10.1137/141000671},
+  urldate = {2024-03-15},
+  abstract = {JuMP is an open-source modeling language that allows users to express a wide range of optimization problems (linear, mixed-integer, quadratic, conic-quadratic, semidefinite, and nonlinear) in a high-level, algebraic syntax. JuMP takes advantage of advanced features of the Julia programming language to offer unique functionality while achieving performance on par with commercial modeling tools for standard tasks. In this work we will provide benchmarks, present the novel aspects of the implementation, and discuss how JuMP can be extended to new problem classes and composed with state-of-the-art tools for visualization and interactivity.},
 }
 
-@software{vaibhav_kumar_dixit_2023_7738525,
-	author = {Vaibhav Kumar Dixit and Christopher Rackauckas},
-	month = mar,
-	publisher = {Zenodo},
-	title = {Optimization.jl: A Unified Optimization Package},
-	version = {v3.12.1},
-	doi = {10.5281/zenodo.7738525},
-  	url = {https://doi.org/10.5281/zenodo.7738525},
-	year = 2023}
-
-@Article{Waechter2006,
-  Title                    = {{O}n the {I}mplementation of an {I}nterior-{P}oint {F}ilter {L}ine-{S}earch {A}lgorithm for {L}arge-{S}cale {N}onlinear {P}rogramming},
-  Author                   = {A. W\"{a}chter and L.T. Biegler},
-  Journal                  = {{M}athematical {P}rogramming},
-  Year                     = {2006},
-  Number                   = {1},
-  Pages                    = {25--57},
-  Volume                   = {106},
-  doi                      = {10.1007/s10107-004-0559-y},
-
-  File                     = {:2006/Waechter2006.pdf:PDF},
-  Journal-iso              = {Math. Program.},
-  Keywords                 = {interior point},
+@article{rackauckas2017,
+  title = {{{DifferentialEquations}}.Jl -- a Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
+  author = {Rackauckas, Christopher and Nie, Qing},
+  year = {2017},
+  journal = {The Journal of Open Research Software},
+  volume = {5},
+  number = {1},
[…]
