### 2024-03-28 de66451
CompatHelper: bump compat for SciMLBase to 2, (keep existing compat)
