+++ Project.toml
@@ -24,7 +24,7 @@ ModelingToolkit = "~8.72.0"
 Optimization = "~3.19.0"
 OrdinaryDiffEq = "6"
 Reexport = "1.2"
-SciMLBase = "~2.7.0"
+SciMLBase = "~2.7.0, 2"
 SciMLSensitivity = "7"
 julia = "1.6"
 
