### 2024-04-29 mathopt/DynamicOED.jl#28 @joshuaeh
Dependency problems in installation
When adding DynamicOED to a new project, with the command `add https://github.com/mathopt/DynamicOED.jl` I get the error:

```
add https://github.com/mathopt/DynamicOED.jl
    Updating git-repo `https://github.com/mathopt/DynamicOED.jl`
    Updating registry at `~/.julia/registries/General.toml`
   Resolving package versions...
ERROR: Unsatisfiable requirements detected for package Logging [56ddb016]:
 Logging [56ddb016] log:
 ├─possible versions are: 1.8.5 or uninstalled (package in sysimage!)
 └─found to have no compatible versions left with SciMLBase [0bca4576]
   └─SciMLBase [0bca4576] log:
     ├─possible versions are: 1.0.0-2.35.0 or uninstalled
     └─restricted to versions 2.7 by DynamicOED [bec7fea8], leaving only versions 2.7.0-2.7.3
       └─DynamicOED [bec7fea8] log:
         ├─possible versions are: 0.1.1 or uninstalled
         └─DynamicOED [bec7fea8] is fixed to version 0.1.1
```

It seems to me this may stem from the SciMLBase version being listed as "~2.7.0" while SciMLBase is only up to version 2.35 as far as I can tell
