### 2024-04-29 #6605 @joshuaeh
Had an issue on installation: https://github.com/mathopt/DynamicOED.jl/issues/28#issue-2269343131
