+++ src/augment.jl
@@ -2,7 +2,7 @@
 """
 $(TYPEDEF)
 
-Indicator that a given variable is a measurement function. 
+Indicator that a given variable is a measurement function.
 
 ```
 @variables w=1.0 [measurement_function=true]
@@ -19,7 +19,7 @@ set_measurement_function(x) = Symbolics.setmetadata(x, MeasurementFunction, true
 """
 $(TYPEDEF)
 
-Indicator that a given variable is a state of the fisher information matrix. 
+Indicator that a given variable is a state of the fisher information matrix.
 
 ```
 @variables F[1:3, 1:3] [fisher_state=true]
@@ -66,8 +66,8 @@ $(TYPEDEF)
 
 Indicator that a given state is subject to a fixed rate.
 Is used for modeling the rate of observation of observed variables and the rate of control for control variables.
-If the provided rate is a `Real`, it is assumed that the resulting time grid is given in fractions of the time unit. 
-If the provided rate is a `Int`, it is assumed that the resulting time grid is divided in `rate` equidistant intervals. 
+If the provided rate is a `Real`, it is assumed that the resulting time grid is given in fractions of the time unit.
+If the provided rate is a `Int`, it is assumed that the resulting time grid is divided in `rate` equidistant intervals.
 
 ```
 @variables y(t) [measurement_rate=0.1] # Create a variable measured every 0.1 t
@@ -112,37 +112,36 @@ end
 
 function construct_jacobians(::MTKBackend,
         sys::ModelingToolkit.AbstractODESystem,
-        p = parameters(sys))
-    eqs = map(x -> x.rhs - x.lhs, equations(sys))
+        p = parameters(sys), stateset = states(sys), equationset = equations(sys))
+    eqs = map(x -> x.rhs - x.lhs, equationset)
     t = ModelingToolkit.get_iv(sys)
     D = Differential(t)
-    fx = ModelingToolkit.jacobian(eqs, states(sys))
-    dfddx = ModelingToolkit.jacobian(eqs, D.(states(sys)))
+    fx = ModelingToolkit.jacobian(eqs, stateset)
+    dfddx = ModelingToolkit.jacobian(eqs, D.(stateset))
     fp = ModelingToolkit.jacobian(eqs, p)
     obs = filter(x -> is_measured(x.lhs), observed(sys))
-    obs = isempty(obs) ? states(sys) : map(x -> x.rhs, obs)
-    hx = ModelingToolkit.jacobian(obs, states(sys))
+    obs = isempty(obs) ? sts : map(x -> x.rhs, obs)
+    hx = ModelingToolkit.jacobian(obs, stateset)
     return dfddx, fx, fp, hx
 end
 
 function build_augmented_system(sys::ModelingToolkit.AbstractODESystem,
-        backend::AbstractAugmentationBackened;
+        backend::AbstractAugmentationBackened; stateset = states(sys),
+        equationset = equations(sys),
         name::Symbol,
         kwargs...)
     T = Float64
     # The set of tuneable parameters
     p = parameters(sys)
-    # The set of controls 
+    # The set of controls
     c = filter(ModelingToolkit.isinput, p)
     # The set of tuneable parameters
     p_tuneable = setdiff(filter(ModelingToolkit.istunable, p), c)
-    # The states
-    x = states(sys)
     # The unknown initial conditions
-    x_ic = eltype(x)[]
+    x_ic = eltype(stateset)[]
     unknown_initial_conditions = Int[]
-    @inbounds for i in axes(x, 1)
-        xi = getindex(x, i)
+    @inbounds for i in axes(stateset, 1)
+        xi = getindex(stateset, i)
         if istunable(xi)
[…]
