+++ docs/src/examples/urethan.md
@@ -1,0 +1,228 @@
+# Design of Experiments for Urethan
+
+In this example we take a look at the reaction of Urethane from the educts Butanol and Isocyanate. The reaction scheme is
+
+```math 
+\begin{aligned}
+A + B &\rightarrow C\\ 
+A + C &\rightleftarrow D\\
+3 A &\rightarrow E
+\end{aligned} 
+```
+with reactants Isocyanate A, Butanol B, Urethane C, Allophanate D and Isocyanurate E.
+
+We start by using [ModelingToolkit.jl](https://github.com/SciML/ModelingToolkit.jl) and DynamicOED.jl to model the system.
+
+```@example urethan
+using DynamicOED
+using ModelingToolkit
+using Optimization, OptimizationMOI, Ipopt
+using Plots
+
+const M    = [ 0.11911, 0.07412, 0.19323, 0.31234, 0.35733, 0.07806 ]
+const rho  = [1095.0, 809.0, 1415.0, 1528.0, 1451.0, 1101.0]
+
+Rg   = 8.314;
+T1   = 363.16;
+
+n_A0, n_B0, n_L0 = 0.1, 0.05, 0.01
+
+# Set up variables
+@variables t
+D = Differential(t)
+@variables h(t)[1:4] [description = "Observed", measurement_rate=10]
+h = collect(h)
+
+@variables n_C(t)=0.0 [description = "Molar numbers for C"]
+@variables n_D(t)=0.0 [description = "Molar numbers for D"]
+@variables n_E(t)=0.0 [description = "Molar numbers for E"]
+@variables n_A(t)=n_A0 [description = "Molar numbers for C"]
+@variables n_B(t)=n_B0 [description = "Molar numbers for D"]
+@variables n_L(t)=n_L0 [description = "Molar numbers for E"]
+
+@parameters begin
+    p1=1.0, [description = "Scaling parameter 1", tunable = true]
+    p2=1.0, [description = "Scaling parameter 2", tunable = true]
+    p3=1.0, [description = "Scaling parameter 3", tunable = true]
+    p4=1.0, [description = "Scaling parameter 4", tunable = false]
+    p5=1.0, [description = "Scaling parameter 5", tunable = false]
+    p6=1.0, [description = "Scaling parameter 6", tunable = false]
+end
+
+# Variables for temperature and feed
+@variables feed1(t)=0 [description = "State for feed 1"]
+@variables feed2(t)=0 [description = "State for feed 2"]
+@variables temperature(t)=373.15 [description = "State for temperature"]
+
+# Controls for temperature and feed
+@parameters begin
+    u1=0.0125, [description = "RHS of feed1", bounds=(0.0,0.0125), input=true, measurement_rate=10]
+    u2=0.0125, [description = "RHS of feed2", bounds=(0.0,0.0125), input=true, measurement_rate=10]
+    u_temp=0.0, [description = "RHS of temperature", bounds=(-15,15), input=false]
+end
+
+
+# Write system of equations
+k_ref1    = p1 * 5.0E-4
+E_a1      = p2 * 35240.0
+k_ref2    = p3 * 8.0E-8
+E_a2      = p4 * 85000.0
+k_ref4    = p5 * 1.0E-8
+E_a4      = p6 * 35000.0
+dH_2      = -17031.0
+K_C2      = 0.17
+
+# Arrhenius equations for the reaction rates
+fac_T = 1.0 / (Rg*temperature) - 1.0 / (Rg*T1);
+k1 = k_ref1 * exp(- E_a1 * fac_T);
+k2 = k_ref2 * exp(- E_a2 * fac_T);
[…]
