### 2024-05-16 #6605 @AlCap23
Hey @diehlpk . Thanks for your efforts!

If needed, I asked two people I know of if they could jump in for a review:

+ @mohamed82008 ; Areas of expertise include OED, Optimization in general and MINLP; Full disclosure: I am working with him on another project within Pumas
+ @baggepinnen ; Areas of expertise include Control, Modeling and Inference; He is also working on ModelingToolkit.jl.

Let me know if I can provide any more help on this.
