### 2024-05-20 #6605 @KBodolai
Overall I'm quite happy with the submission, the text in the paper reads well and the software looks solid!

A couple of small things that I noticed while reviewing, preventing me from ticking the last two boxes:

1. I can't seem to reproduce the examples other than the example in the home section, more info in mathopt/DynamicOED.jl#30.
2. (A bit more of a nitpick) As for the community guidelines tickbox, I've not seen any guidelines in the documentation referring to contributions. (issues and support can easily be reported through the gh issues). @AlCap23 , could it be possible to add a contributing.md file, or anything in the docs clarifying this?
3. There seems to be an error rendering the figure in the paper. I'm not entirely sure, but it may be because it's a pdf.
