### 2024-05-21 0b9cd4d
CompatHelper: bump compat for ModelingToolkit to 9, (keep existing compat)
