+++ Project.toml
@@ -20,7 +20,7 @@ Aqua = "0.7"
 CommonSolve = "0.2"
 ComponentArrays = "0.15"
 DocStringExtensions = "0.9"
-ModelingToolkit = "~8.72.0"
+ModelingToolkit = "~8.72.0, 9"
 Optimization = "~3.19.0"
 OrdinaryDiffEq = "6"
 Reexport = "1.2"
