### 2024-05-21 mathopt/DynamicOED.jl#30 @AlCap23
Thanks @KBodolai !

Yes, this is a missing `Num[]` in the `ConstraintSystem` and it seems I forgot to update the docs accordingly 😅 .
