+++ docs/src/examples/1D.md
@@ -69,7 +69,7 @@ constraint_equations = [
 ]
 
 @named constraint_system = ConstraintsSystem(
-    constraint_equations, collect(optimization_variables), []
+    constraint_equations, collect(optimization_variables), Num[]
 )
 nothing # hide
 ```
