### 2024-05-21 #6605 @AlCap23
Thanks for the review and the approval!

> 1. I can't seem to reproduce the examples other than the example in the home section, more info in [MethodError in examples. mathopt/DynamicOED.jl#30](https://github.com/mathopt/DynamicOED.jl/issues/30).
> 2. (A bit more of a nitpick) As for the community guidelines tickbox, I've not seen any guidelines in the documentation referring to contributions. (issues and support can easily be reported through the gh issues). @AlCap23 , could it be possible to add a contributing.md file, or anything in the docs clarifying this?
> 3. There seems to be an error rendering the figure in the paper. I'm not entirely sure, but it may be because it's a pdf. I did a quick test with a screenshot and it renders well saved as a png, so I'd suggest exporting `paper/figures/lotka.pdf` as a png and replacing it. (I'd do a pull request with it, but it's probably best to export it directly as a png to preserve image quality)

Should be handled via https://github.com/mathopt/DynamicOED.jl/pull/31
Instead of a seperate contribution guide, we choose to follow the ColPrac given by SciML ( as we are already closely related to this part of Julias Ecosystem ).
