### 2024-05-21 fee00ec
added missing Num in constraints (#31)
* added missing Num in constraints

* added plot as png

* added information on contributions
