+++ README.md
@@ -3,6 +3,7 @@
 [![Latest Docs](https://img.shields.io/badge/docs-dev-blue.svg)](https://mathopt.github.io/DynamicOED.jl/dev/)
 [![Aqua QA](https://raw.githubusercontent.com/JuliaTesting/Aqua.jl/master/badge.svg)](https://github.com/JuliaTesting/Aqua.jl) [![SciML Code Style](https://img.shields.io/static/v1?label=code%20style&message=SciML&color=9558b2&labelColor=389826)](https://github.com/SciML/SciMLStyle) ![Lifecycle:Experimental](https://img.shields.io/badge/Lifecycle-Experimental-339999) 
 [![Build Status](https://github.com/mathopt/DynamicOED.jl/actions/workflows/CI.yml/badge.svg?branch=main)](https://github.com/mathopt/DynamicOED.jl/actions/workflows/CI.yml?query=branch%3Amain)
+[![ColPrac: Contributor's Guide on Collaborative Practices for Community Packages](https://img.shields.io/badge/ColPrac-Contributor's%20Guide-blueviolet)](https://github.com/SciML/ColPrac)
 
 Repository for optimal experimental design for differential equations using optimal control.
 
