### 2024-05-21 mathopt/DynamicOED.jl#30 @AlCap23
It seems there is an error in the docs build. The PR #31 updated the docs, but it did not get published.
