### 2024-05-21 #6605 @KBodolai
fantastic, thanks for checking so quickly! I've verified that the examples are now working and provide the expected results.

Good to hear about the separate contribution guide, it could be a good idea to mention it in the docs though (thinking of people that may be newer to the Julia ecosystem).
