### 2024-06-10 #6605 @diehlpk
Hi @AlCap23 I took a look at the paper as editor and here are my comments:

-  Figure 1: Please add a unit to the `Time` is it seconds or milliseconds?
-  Line 55 add a space between ModelingToolkit.jl and the reference.
