### 2024-06-18 #6605 @AlCap23
> @AlCap23 if you could address these comments, we could proceed with the paper.

@diehlpk I've added the space. Unfortunately, time has no unit in given that it is from our viewpoint just a "scaling" factor ( see also the [original publication](https://mathopt.de/publications/Sager2013.pdf) ), so I would just leave it blank or add `days` if strictly necessary.
