### 2024-06-18 ecc3a1f
Update Project.toml
Up versioning ( for zenodo )
