### 2024-06-18 #6605 @AlCap23
@diehlpk We archived the version v0.1.2 on Zenodo and are good to go 👍 License, Orcid IDs and authors check out.

One minor comment based on my copy-pasta mistake : There is a `1` in the title of the paper hidden, would you be so kind to remove it? Thanks!
