### 2024-06-18 #6605 @diehlpk
@AlCap23 `One minor comment based on my copy-pasta mistake : There is a 1 in the title of the paper hidden, would you be so kind to remove it? Thanks!`

Not sure where this is happening. I think you should be able to fix that too.
