### 2024-06-18 #6605 @diehlpk
@AlCap23 if you log in to Zenodo, you should be able to edit the title and other meta data. See the screenshot below with one from my projects. I do not have access to your project to change that.
