### 2024-06-18 #6605 @AlCap23
I've changed the title to

DynamicOED.jl: A Julia package for solving optimum1 experimental design problems

without the typo :) . Does this work for you @diehlpk ?
