### 2024-06-18 #6605 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
