### 2024-06-18 25c8179
small changes in paper.md
