### 2024-06-18 3407d96
minor changes in bibliography.bib
