+++ paper/bibliography.bib
@@ -1,5 +1,5 @@
 @article{bezanson2017julia,
-  title = {Julia: {{A Fresh Approach}} to {{Numerical Computing}}},
+  title = {Julia: A Fresh Approach to Numerical Computing},
   shorttitle = {Julia},
   author = {Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
   year = {2017},
@@ -16,7 +16,7 @@
 }
 
 @article{rackauckas2017,
-  title = {{{DifferentialEquations}}.Jl -- a Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
+  title = {{DifferentialEquations.jl} -- a Performant and Feature-Rich Ecosystem for Solving Differential Equations in {J}ulia},
   author = {Rackauckas, Christopher and Nie, Qing},
   year = {2017},
   journal = {The Journal of Open Research Software},
@@ -27,7 +27,7 @@
 }
 
 @misc{vaibhav_kumar_dixit_2023_7738525,
-  title = {Optimization.Jl: {{A}} Unified Optimization Package},
+  title = {{Optimization.jl}: A Unified Optimization Package},
   author = {Dixit, Vaibhav Kumar and Rackauckas, Christopher},
   year = {2023},
   month = mar,
@@ -36,7 +36,7 @@
 }
 
 @phdthesis{Koerkel2002,
-  title = {Numerische {{Methoden}} F{\"u}r {{Optimale Versuchsplanungsprobleme}} Bei Nichtlinearen {{DAE-Modellen}}},
+  title = {Numerische Methoden F{\"u}r Optimale Versuchsplanungsprobleme Bei Nichtlinearen {DAE-Modellen}},
   author = {K{\"o}rkel, S.},
   year = {2002},
   address = {{Heidelberg}},
@@ -46,7 +46,7 @@
 }
 
 @inproceedings{juniper,
-  title = {Juniper: {{An Open-Source Nonlinear Branch-and-Bound Solver}} in {{Julia}}},
+  title = {Juniper: An Open-Source Nonlinear Branch-and-Bound Solver in {J}ulia},
   shorttitle = {Juniper},
   booktitle = {Integration of {{Constraint Programming}}, {{Artificial Intelligence}}, and {{Operations Research}}},
   author = {Kr{\"o}ger, Ole and Coffrin, Carleton and Hijazi, Hassan and Nagarajan, Harsha},
@@ -62,7 +62,7 @@
 }
 
 @article{Li2000SensitivityAnalysisDifferential,
-  title = {Sensitivity Analysis of {{Differential}}--{{Algebraic}} Equations: {{A}} Comparison of Methods on a Special Problem},
+  title = {Sensitivity Analysis of Differential--Algebraic Equations: A Comparison of Methods on a Special Problem},
   shorttitle = {Sensitivity Analysis of {{Differential}}--{{Algebraic}} Equations},
   author = {Li, Shengtai and Petzold, Linda and Zhu, Wenjie},
   year = {2000},
@@ -80,7 +80,7 @@
 }
 
 @misc{ma2021modelingtoolkit,
-  title = {{{ModelingToolkit}}: {{A Composable Graph Transformation System For Equation-Based Modeling}}},
+  title = {{ModelingToolkit}: A Composable Graph Transformation System For Equation-Based Modeling},
   shorttitle = {{{ModelingToolkit}}},
   author = {Ma, Yingbo and Gowda, Shashi and Anantharaman, Ranjan and Laughman, Chris and Shah, Viral and Rackauckas, Chris},
   year = {2022},
@@ -96,7 +96,7 @@
 }
 
 @article{Sager2013,
[…]
