### 2024-06-18 #6605 @danielskatz
@AlCap23 - as the track editor, I've proofread the paper, and suggest the changes in https://github.com/mathopt/DynamicOED.jl/pull/33  Please merge this, or let me know what you disagree with, then we can continue to acceptance and publication.
