### 2024-06-19 #6605 @danielskatz
Congratulations to @AlCap23 (Julius Martensen) and co-authors on your publication!!

JOSS depends on volunteers and we couldn't do this with you
